Move currently active tab to leftmost position and active tab next to it
This seems to be the only way to make sure that a tab receiving a
message is visible to the user. If we just move the active tab, it may
still be out of sight if there are so many tabs that not all fit the the
same time to the tab bar.